# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finance API",
        "description": "Personal Finance Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Retrieve user's finance summary",
                "operationId": "455d87bd34098d5282aff2493c6c6fce",
                "responses": {
                    "200": {
                        "description": "Dashboard summary returned"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/advanced": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Retrieve advanced financial metrics",
                "operationId": "729fd4b515657d8e7566c2b8e2de4668",
                "responses": {
                    "200": {
                        "description": "Advanced dashboard metrics returned"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}